requirements.txt
meshtastic
pyserial
PyPubSub
